(subtree) fix: update js script to ts (#3543)
